- Navigaiton3 With KoIn

+++ domain/src/main/java/com/app/domain/model/Coin.kt
@@ -9,13 +9,13 @@ data class CryptoCurrency(
     val priceChangePercentage24h: Double,
 
     // Detail screen fields
-    val marketCap: Long,
+    val marketCap: Double,
     val marketCapRank: Int,
-    val totalVolume: Long,
+    val totalVolume: Double,
     val high24h: Double,
     val low24h: Double,
     val priceChange24h: Double,
-    val marketCapChange24h: Long,
+    val marketCapChange24h: Double,
     val marketCapChangePercentage24h: Double,
     val circulatingSupply: Double,
     val totalSupply: Double?,
